chore: moving date to be a parameter on client
